[WORKING] add coursesShow - moved old controllers to pkg 'old' - automatically removed unused imports
