+++ persistence/src/main/java/ar/edu/itba/paw/persistence/UserHibernateDao.java
@@ -7,13 +7,15 @@ import ar.edu.itba.paw.shared.Result;
 import ar.edu.itba.paw.shared.UserFilter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
-import javax.persistence.criteria.*;
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Predicate;
+import javax.persistence.criteria.Root;
 import javax.persistence.metamodel.EntityType;
 import java.util.ArrayList;
 import java.util.Collections;
