+++ webapp/src/main/java/ar/edu/itba/paw/webapp/config/WebAuthConfig.java
@@ -9,8 +9,6 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
-import java.util.concurrent.TimeUnit;
-
 @Configuration
 @EnableWebSecurity
 @ComponentScan(
