+++ webapp/src/main/java/ar/edu/itba/paw/webapp/controllers/old/HTTPErrorsController.java
@@ -1,4 +1,4 @@
-package ar.edu.itba.paw.webapp.controllers;
+package ar.edu.itba.paw.webapp.controllers.old;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
